3단계 리팩토링 - 불필요한 일급컬렉션 간소화
