+++ src/main/java/ladder/Ladder.java
@@ -6,22 +6,22 @@ import ladder.generator.PointGenerator;
 
 public class Ladder {
 
-    private final Lines lines;
+    private final List<Line> lines;
 
     public Ladder(PointGenerator pointGenerator, final Size size) {
         List<Line> lines = new ArrayList<>();
         for (int i = 0; i < size.getHeight(); i++) {
-            lines.add(new Line(pointGenerator, size.getWidth()));
+            lines.add(new Line(pointGenerator.generate(size.getWidth())));
         }
-        this.lines = Lines.of(lines);
+        this.lines = lines;
     }
 
-    public Lines getLines() {
+    public List<Line> getLines() {
         return lines;
     }
 
     public int play(int currentIndex) {
-        for (Line line : lines.getLines()) {
+        for (Line line : lines) {
             currentIndex = line.move(currentIndex);
         }
         return currentIndex;
